+++ NEWS.md
@@ -1,3 +1,8 @@
+# zoltr 0.5.9
+- Fixed `job_data()` to return `units` column as character, not numeric.
+- Added `units` arg to `do_zoltar_query()` to match `submit_query()`.
+
+
 # zoltr 0.5.8
 - Changed `busy_poll_job()` to check for the new TIMEOUT status.